Renamed old ApplicationBar.tsx and removed it from use. Renamed Footer to be ApplicationBar.tsx.
