+++ src/App.tsx
@@ -1,6 +1,5 @@
 import { Sidebar } from './components/Sidebar';
 import { MainView } from './components/MainView';
-import { ApplicationBar } from './components/ApplicationBar';
 import { Outlet, useNavigation } from 'react-router-dom';
 import { useEffect } from 'react';
 import { getSchemaToStore } from './utils/storeUtils';
@@ -16,7 +15,6 @@ function App() {
         <div className="App w-full flex relative">
             <Sidebar />
             <MainView>
-                <ApplicationBar />
                 {navigation.state === 'loading' && (
                     <p className={`px-8 pt-4 text-2xl`}>Loading page...</p>
                 )}
