+++ src/components/ApplicationBar.tsx
@@ -1,16 +1,31 @@
-export const ApplicationBar = () => {
-    return (
+import { createPortal } from 'react-dom';
+import { useEffect, useState } from 'react';
+
+type footerProps = {
+    children: React.ReactNode;
+    className?: string;
+};
+
+export const ApplicationBar = (props: footerProps) => {
+    const [mainElement, setMainElement] = useState<HTMLElement | null>(null);
+
+    useEffect(() => {
+        const mainEl = document.querySelector('main');
+        setMainElement(mainEl);
+    }, []);
+
+    if (!mainElement) {
+        return null;
+    }
+
+    return createPortal(
         <div
-            className={`toolbar bg-white border-b-2 border-ad-blue-600 flex justify-between px-4`}
+            className={`toolbar bg-white border-t-2 border-ad-blue-600 fixed bottom-px px-4 w-full ${
+                props.className && props.className
+            }`}
         >
-            <div className={`page-actions`}>
-                <span>⬅️</span>
-                <span>➡️</span>
-            </div>
-            <div className="current-page">Kotisivu</div>
-            <div className="windows-actions">
-                <span>❔</span>
-            </div>
-        </div>
+            {props.children}
+        </div>,
+        mainElement,
     );
 };
