+++ src/test/React.ApplicationBar.test.tsx
@@ -1,15 +1,15 @@
 import { render, screen } from '@testing-library/react';
-import { ApplicationBar } from '../components/ApplicationBar';
+import { ApplicationBarOld } from '../components/ApplicationBarOld';
 import '@testing-library/jest-dom';
 
-describe('Register ApplicationBar', () => {
+describe('Register ApplicationBarOld', () => {
     it('should render the element', () => {
-        render(<ApplicationBar />);
+        render(<ApplicationBarOld />);
         expect(screen.getByText(/Kotisivu/)).toBeInTheDocument();
     });
 
     it('should have a help icon', () => {
-        render(<ApplicationBar />);
+        render(<ApplicationBarOld />);
         expect(screen.getByText(/❔/)).toBeInTheDocument();
     });
 });
